move `authorization_code` logic
