+++ src/service/src/lib.rs
@@ -2,10 +2,11 @@
 
 #![forbid(unsafe_code)]
 
-pub mod auth;
+pub mod auth_svc;
 pub mod client;
 pub mod encryption;
 pub mod login_delay;
+pub mod oidc;
 pub mod password_reset;
 pub mod suspicious_request_block;
 pub mod token_set;
