+++ src/service/src/oidc/mod.rs
@@ -1,0 +1,1 @@
+pub mod authorization_code;
